experiments/DualViewHair/requirements.txt
# DualViewHair Dependencies

# Core ML frameworks
torch>=1.10.0
torchvision>=0.11.0

# Computer vision
Pillow>=8.0.0

# Data processing
numpy>=1.20.0
tqdm>=4.62.0

# Visualization (for qualitative test)
matplotlib>=3.5.0

# Configuration and logging
PyYAML>=5.4.0
tensorboard>=2.7.0
tqdm>=4.62.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Jupyter notebooks (for analysis)
jupyter>=1.0.0
ipywidgets>=7.6.0

# Development tools
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0

# Optional: Advanced features
# faiss-cpu>=1.7.0  # For fast similarity search
# umap-learn>=0.5.0  # For embedding visualization
# albumentations>=1.1.0  # Advanced augmentations
